requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# PyTorch and Transformers for CLIP
torch>=2.6.0
torchvision>=0.17.0
transformers>=4.36.2

# Image processing
Pillow>=10.2.0

# Data validation
pydantic>=2.5.3
pydantic-settings>=2.1.0

# HTTP client
httpx>=0.26.0

# Utilities
numpy>=1.26.3
